requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.0+cpu
torchvision==0.19.0+cpu